+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 - Chevre返金処理をクレジットカード以外の決済方法に対応
 - オファーの適用ムビチケ条件の決済方法として、appliesToMovieTicket.serviceOutput.typeOfを参照するように変更
+- 注文取引確定時のムビチケ系統決済に対する検証処理を、利用可能なムビチケ系統決済方法タイプに対して動的に実行するように調整
 
 ### Deprecated